src/test/java/com/leetcode/dp/hard/KInversePairsArray_629Test.java
package com.leetcode.dp.hard;

import org.junit.jupiter.api.extension.ExtensionContext;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.ArgumentsProvider;
import org.junit.jupiter.params.provider.ArgumentsSource;

import java.util.stream.Stream;

import static org.assertj.core.api.Assertions.assertThat;
import static org.junit.jupiter.params.provider.Arguments.arguments;

class KInversePairsArray_629Test {

  private final KInversePairsArray_629 solution = new KInversePairsArray_629();

  private static class TestArgumentsProvider implements ArgumentsProvider {
    @Override
    public Stream<? extends Arguments> provideArguments(ExtensionContext context) {
      return Stream.of(arguments(3, 0, 1), arguments(3, 1, 2));
    }
  }

  @ParameterizedTest
  @ArgumentsSource(TestArgumentsProvider.class)
  void kInversePairs(int n, int k, int expectedResult) {
    int actualResult = solution.kInversePairs(n, k);

    assertThat(actualResult).isEqualTo(expectedResult);
  }
}
